+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0-alpha.13] - 2026-06-12
+
 ### Fixed
 
 - migration: classify `RunnerError` outcomes for `djogi migrations apply` so